+++ physical_ai_server/physical_ai_server/training/training_manager.py
@@ -122,8 +122,10 @@ class TrainingManager:
         current_training_status.training_info = self.training_info
         if self.trainer:
             current_training_status.current_step = self.trainer.get_current_step()
+            current_training_status.current_loss = self.trainer.get_current_loss()
         else:
             current_training_status.current_step = 0
+            current_training_status.current_loss = float('nan')
         return current_training_status
 
     def train(self):